Consolidate logic up to magic square games layer
