+++ src/math-games/magic-square-games/fifteen-game/FifteenGameRoutes.js
@@ -1,21 +1,27 @@
 import React, { lazy, useState, useEffect } from 'react'
 import { Routes, Route } from 'react-router-dom';
-
 import {  Container,  Box } from '@mui/material';
 
-import { generatePositionToOutcomeMap } from "./logic/gameLogic";
+import { generatePositionToOutcomeMap } from "../solution";
+
 
 const Welcome = lazy(() => import('./pages/Welcome'))
 // const PlayWithCoach = lazy(() => import('./pages/PlayWithCoach'))
 // const PlayVsFriend = lazy(() => import('./pages/PlayVsFriend')) 
 const PlayVsBot = lazy(() => import('./pages/PlayVsBot')) 
 
+// Game Logic
+
 export default function FifteenGameRoutes(props) {
   const [outcomeMap, setOutcomeMap] = useState(null)
+  
   useEffect(() => {
-    let map = generatePositionToOutcomeMap()
-    setOutcomeMap(map)
-    console.log("Outcome Map Generated");
+    const generateOutcomeMapAsync = async () => {
+      const map = await generatePositionToOutcomeMap()
+      setOutcomeMap(map)
+      console.log("Outcome Map Generated");
+    }
+    generateOutcomeMapAsync()
   }, [])
 
   return (
