+++ src/math-games/magic-square-games/fifteen-game/botLogic.js
@@ -1,8 +1,8 @@
 // X always goes first! X and O will be shorthand for the player who goes First and the player who goes second.
 // PlayerOne and PlayerTwo are NOT markers of who goes first or second in any given game!
 // In the 15 game, PlayerOne and PlayerTWo are distinguished by dark-blue and light-blue
-import { intersect } from "../../_helpers/probability";
-import { sumsOfTwo, complementOf, gameOver, xNumbers, oNumbers, nextPlayer, availableNumbers } from "../solution";
+import { sumsOfTwo, complementOf, intersect, gameOver, xNumbers, oNumbers, nextPlayer, availableNumbers } from "./gameLogic";
+// const outcomeMap = generatePositionToOutcomeMap()
 
 
 //////////////////////////////////////////////////////////////
