+++ src/math-games/magic-square-games/tic-tac-toe/components/panels/Commentary.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { nextPlayer, status, getParent } from "../../../magicSquareGameSolution";
+import { nextPlayer, status, getParent } from "../../../solution";
 import { winningMoves, urgentDefensiveMoves, doubleAttackingMoves } from "../../botLogic";
