+++ src/math-games/magic-square-games/tic-tac-toe/pages/PlayWithCoach.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 // import { useOutletContext } from "react-router-dom";
 
-import { gameOver } from "../../magicSquareGameSolution";
+import { gameOver } from "../../solution";
 import Board from "../components/Board";
 import CoachPanel from "../components/panels/CoachPanel";
